+++ examples/terachem_opt.py
@@ -30,7 +30,7 @@ try:
     # prog_output is instance of Results
     prog_output = compute("terachem", prog_input, collect_files=True)
 except exceptions.QCOPBaseError as e:
-    prog_output = e.program_output
+    prog_output = e.results
     print(prog_output.stdout)  # or output.pstdout for short
     print(f"Success: {prog_output.success}")  # False
     print(prog_output.input_data)  # Input data used to generate the calculation